+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Graphql exception `BackendValidation`.
+- `returnsOnException` method to `Query` test helper class.
 
 ## [0.26.0] - 2024-04-04