+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.27.0] - 2024-05-31
+
 ### Added
 - Graphql exception `BackendValidation`.
 - `returnsOnException` method to `Query` test helper class.